Move all the specimen writers and SpecimenWebPartFactory to specimen-main
Move BaseSpecimenPivotTable to study-api
