+++ study/api-src/org/labkey/api/study/writer/BaseStudyWriter.java
@@ -1,6 +1,5 @@
 package org.labkey.api.study.writer;
 
-import org.labkey.api.specimen.writer.SimpleStudyExportContext;
 import org.labkey.api.study.Study;
 import org.labkey.api.writer.Writer;
