+++ study/api-src/org/labkey/api/study/writer/SimpleStudyWriter.java
@@ -1,0 +1,7 @@
+package org.labkey.api.study.writer;
+
+import org.labkey.api.study.Study;
+
+public interface SimpleStudyWriter extends BaseStudyWriter<Study, SimpleStudyExportContext>
+{
+}
