+++ study/src/org/labkey/study/query/StudyQuerySchema.java
@@ -23,8 +23,6 @@ import org.labkey.api.action.SpringActionController;
 import org.labkey.api.data.ColumnInfo;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.ContainerFilter;
-import org.labkey.api.data.DbSchema;
-import org.labkey.api.data.DbSchemaType;
 import org.labkey.api.data.SQLFragment;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.exp.property.Domain;
@@ -158,7 +156,7 @@ public class StudyQuerySchema extends UserSchema
     private Map<Integer, List<Double>> _datasetSequenceMap;
     public static final String STUDY_DATA_TABLE_NAME = "StudyData";
     public static final String QCSTATE_TABLE_NAME = "QCState";
-    private Set<String> _tableNames;
+    protected Set<String> _tableNames;
 
     private ParticipantGroup _sessionParticipantGroup;
 
@@ -181,7 +179,7 @@ public class StudyQuerySchema extends UserSchema
     /**
      * This c-tor is for nested study schemas
      */
-    private StudyQuerySchema(SchemaKey path, String description, @Nullable StudyImpl study, Container c, User user, boolean mustCheckPermissions)
+    protected StudyQuerySchema(SchemaKey path, String description, @Nullable StudyImpl study, Container c, User user, boolean mustCheckPermissions)
     {
         super(path, description, user, c, StudySchema.getInstance().getSchema(), null);
         _study = study;
@@ -230,11 +228,6 @@ public class StudyQuerySchema extends UserSchema
         if (StringUtils.equalsIgnoreCase("Specimens",name))
             return new SpecimenSchema(this);
         return super.getSchema(name);
-    }
-
-    private DbSchema getStudyDesignSchema()
-    {
-        return DbSchema.get(STUDY_DESIGN_SCHEMA_NAME, DbSchemaType.Provisioned);
     }
 
     public String getSubjectColumnName()
@@ -1266,71 +1259,6 @@ public class StudyQuerySchema extends UserSchema
 
                 _tableNames = Collections.unmodifiableSet(names);
             }
-            return _tableNames;
-        }
-    }
-
-    private class SpecimenSchema extends StudyQuerySchema
-    {
-        final StudyQuerySchema _parentSchema;
-
-        SpecimenSchema(StudyQuerySchema parent)
-        {
-            super(new SchemaKey(parent.getSchemaPath(), "Specimens"), "Specimen repository", parent.getStudy(), parent.getContainer(), parent.getUser(), parent._mustCheckPermissions);
-            _parentSchema = parent;
-            setSessionParticipantGroup(parent.getSessionParticipantGroup());
-        }
-
-        @Override
-        public Set<String> getSubSchemaNames()
-        {
-            return Collections.emptySet();
-        }
-
-        @Override
-        public QuerySchema getSchema(String name)
-        {
-            return _parentSchema.getSchema(name);
-        }
-
-        @Override
-        public Set<String> getTableNames()
-        {
-            if (_tableNames == null)
-            {
-                Set<String> names = new LinkedHashSet<>();
-
-                if (_study != null)
-                {
-                    StudyService studyService = StudyService.get();
-                    if (null == studyService)
-                        throw new IllegalStateException("No StudyService!");
-
-                    names.add(LOCATION_TABLE_NAME);
-                    names.add(SPECIMEN_EVENT_TABLE_NAME);
-                    names.add(SPECIMEN_DETAIL_TABLE_NAME);
-                    names.add(SPECIMEN_SUMMARY_TABLE_NAME);
-                    names.add("SpecimenVialCount");
-                    names.add(SIMPLE_SPECIMEN_TABLE_NAME);
-                    names.add("SpecimenRequest");
-                    names.add("SpecimenRequestStatus");
-                    names.add("VialRequest");
-                    names.add(SPECIMEN_ADDITIVE_TABLE_NAME);
-                    names.add(SPECIMEN_DERIVATIVE_TABLE_NAME);
-                    names.add(SPECIMEN_PRIMARY_TYPE_TABLE_NAME);
-                    names.add("SpecimenComment");
-
-                    // CONSIDER: show under queries instead of tables?
-                    // specimen report pivots
-                    names.add(SpecimenPivotByPrimaryType.PIVOT_BY_PRIMARY_TYPE);
-                    names.add(SpecimenPivotByDerivativeType.PIVOT_BY_DERIVATIVE_TYPE);
-                    names.add(SpecimenPivotByRequestingLocation.PIVOT_BY_REQUESTING_LOCATION);
-
-                    names.add(LOCATION_SPECIMEN_LIST_TABLE_NAME);
-                }
-                _tableNames = Collections.unmodifiableSet(names);
-            }
-
             return _tableNames;
         }
     }
